Extract no retries RPC handler to separate class
